requirements.txt
fastapi~=0.112.2
uvicorn==0.35.0
aiofiles~=22.1.0
aiosqlite
torch~=2.2.2
torchvision~=0.17.2
pillow~=11.3.0
sqlalchemy~=2.0.41
alembic==1.16.4
werkzeug==3.1.3
scikit-learn~=1.7.0
psutil~=5.9.0
alembic~=1.16.4
python-dotenv~=1.1.1